backend_end/requirements.txt
# ── Web framework ────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# ── Config ───────────────────────────────────────────────────────
pydantic-settings==2.3.4

# ── HTTP client ──────────────────────────────────────────────────
httpx==0.27.0

# ── Database ─────────────────────────────────────────────────────
sqlalchemy==2.0.31
asyncpg==0.29.0
alembic==1.13.2
pgvector==0.3.2

# ── Cache ────────────────────────────────────────────────────────
redis==5.0.8

# ── Web tools ────────────────────────────────────────────────────
duckduckgo-search==8.1.1
beautifulsoup4==4.12.3
lxml>=5.3.0

# ── File processing ──────────────────────────────────────────────
pdfplumber==0.11.2
python-docx==1.1.2
python-pptx==1.0.2

# ── TTS ──────────────────────────────────────────────────────────
edge-tts==6.1.12

# ── Local LLM (router) ───────────────────────────────────────────
llama-cpp-python==0.3.4

# ── Utilities ────────────────────────────────────────────────────
tenacity==8.5.0

# ── Testing ──────────────────────────────────────────────────────
pytest==8.3.2
pytest-asyncio==0.23.8
anyio==4.4.0
httpx==0.27.0          # already above, for AsyncClient in tests
pytest-cov==5.0.0

# ── Linting (optional, not installed in Docker prod) ─────────────
# ruff==0.5.7
